20589: Migrate case client methods (#227)
- Move general case operations from the concrete client into the
abstract base client
- Some type corrections and migrations
- Resolve circular dependencies

Note: This is just a partial step of the full migration to keep line
changes manageable and is not intended to be fully functional at this
time.
