+++ howso/direct/core.py
@@ -843,22 +843,6 @@ class HowsoCore:
             "condition_session": condition_session,
         })
 
-    def get_num_training_cases(self, trainee_id: str) -> Dict:
-        """
-        Return the number of trained cases in the model.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-
-        Returns
-        -------
-        dict
-            A dictionary containing the key "count".
-        """
-        return self._execute(trainee_id, "get_num_training_cases", {})
-
     def get_auto_ablation_params(self, trainee_id: str):
         """Get trainee parameters for auto-ablation set by :meth:`set_auto_ablation_params`."""
         return self._execute(
@@ -1117,38 +1101,6 @@ class HowsoCore:
         """
         return self._execute(trainee_id, "get_metadata", {})
 
-    def retrieve_extreme_cases_for_feature(
-        self,
-        trainee_id: str,
-        num: int,
-        sort_feature: str,
-        features: Optional[Iterable[str]] = None
-    ) -> Dict:
-        """
-        Gets the extreme cases of a Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        num : int
-            The number of cases to get.
-        sort_feature : str
-            The feature name by which extreme cases are sorted by.
-        features: iterable of str, optional
-            An iterable of feature names to use when getting extreme cases.
-
-        Returns
-        -------
-        dict
-            A dictionary of keys 'cases' and 'features'.
-        """
-        return self._execute(trainee_id, "retrieve_extreme_cases_for_feature", {
-            "features": features,
-            "sort_feature": sort_feature,
-            "num": num,
-        })
-
     def train(
         self,
         trainee_id: str,
@@ -1219,42 +1171,6 @@ class HowsoCore:
             "train_weights_only": train_weights_only,
         })
 
-    def impute(
-        self,
-        trainee_id: str,
-        *,
-        batch_size: int = 1,
-        features: Optional[Iterable[str]] = None,
-        features_to_impute: Optional[Iterable[str]] = None,
-        session: Optional[str] = None
-    ) -> None:
-        """
-        Impute, or fill in the missing values, for the specified features.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        batch_size : int, default 1
-            Larger batch size will increase accuracy and decrease speed.
-            Batch size indicates how many rows to fill before recomputing
-            conviction.
-        features : iterable of str, optional
-            An iterable of feature names to use for imputation. If not
-            specified, all features will be used imputed.
-        features_to_impute : iterable of str, optional
-            An iterable of feature names to impute. If not specified, features
-            will be used (see above).
-        session : str, optional
-            The identifier of the Trainee session to associate the edit with.
-        """
-        return self._execute(trainee_id, "impute", {
-            "features": features,
-            "features_to_impute": features_to_impute,
-            "session": session,
-            "batch_size": batch_size,
-        })
-
     def clear_imputed_session(
         self,
         trainee_id: str,
@@ -1279,65 +1195,6 @@ class HowsoCore:
         return self._execute(trainee_id, "clear_imputed_session", {
             "session": session,
             "impute_session": impute_session,
-        })
-
-    def get_cases(
-        self,
-        trainee_id: str,
-        session: Optional[str] = None,
-        *,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        indicate_imputed: bool = False,
-        features: Optional[Iterable[str]] = None,
-        condition: Optional[Dict] = None,
-        num_cases: Optional[int] = None,
-        precision: Optional[Literal["exact", "similar"]] = None
-    ) -> Dict:
-        """
-        Retrieve cases from a Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        session : str, optional
-            The session identifier to retrieve cases for, in their trained
-            order.
-        case_indices : iterable of sequence of str, int, optional
-            Iterable of Sequences, of session id and index, where index is the
-            original 0-based index of the case as it was trained into the
-            session. If specified, returns only these cases and ignores the
-            session parameter.
-        indicate_imputed : bool, default False
-            If set, an additional value will be appended to the cases
-            indicating if the case was imputed.
-        features : iterable of str, optional
-            A list of feature names to return values for in leu of all
-            default features.
-        condition : dict, optional
-            The condition map to select the cases to retrieve that meet all the
-            provided conditions.
-        num_cases : int, default None
-            The maximum amount of cases to retrieve. If not specified, the limit
-            will be k cases if precision is "similar", or no limit if precision
-            is "exact".
-        precision : {"exact", "similar}, optional
-            The precision to use when retrieving the cases via condition.
-            If not provided, "exact" will be used.
-
-        Returns
-        -------
-        dict
-            A dictionary containing keys 'features' and 'cases'.
-        """
-        return self._execute(trainee_id, "get_cases", {
-            "features": features,
-            "session": session,
-            "case_indices": case_indices,
-            "indicate_imputed": indicate_imputed,
-            "condition": condition,
-            "num_cases": num_cases,
-            "precision": precision,
         })
 
     def append_to_series_store(
@@ -2609,22 +2466,6 @@ class HowsoCore:
             "weight_feature": weight_feature,
         })
 
-    def set_random_seed(self, trainee_id: str, seed: Union[int, float, str]
-                        ) -> None:
-        """
-        Sets the random seed for the Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        seed: int or float or str
-            The random seed.
-        """
-        return self._execute(trainee_id, "set_random_seed", {
-            "seed": seed,
-        })
-
     def get_internal_parameters(
         self,
         trainee_id: str,
@@ -2674,206 +2515,6 @@ class HowsoCore:
             "mode": mode,
             "weight_feature": weight_feature,
         })
-
-    def move_cases(
-        self,
-        trainee_id: str,
-        num_cases: int = 1,
-        *,
-        case_indices: Optional[Iterable[Tuple[str, int]]] = None,
-        condition: Optional[Dict] = None,
-        condition_session: Optional[str] = None,
-        distribute_weight_feature: Optional[str] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-        preserve_session_data: bool = False,
-        session: Optional[str] = None,
-        source_id: Optional[str] = None,
-        source_name_path: Optional[List[str]] = None,
-        target_name_path: Optional[List[str]] = None,
-        target_id: Optional[str] = None
-    ) -> Dict:
-        """
-        Moves cases from one trainee to another in the hierarchy.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee doing the moving.
-        num_cases : int
-            The number of cases to move; minimum 1 case must be moved.
-            Ignored if case_indices is specified.
-        case_indices : list of tuples
-            A list of tuples containing session ID and session training index
-            for each case to be removed.
-        condition : dict, optional
-            The condition map to select the cases to move that meet all the
-            provided conditions. Ignored if case_indices is specified.
-        condition_session : str, optional
-            If specified, ignores the condition and operates on cases for
-            the specified session id. Ignored if case_indices is specified.
-        precision : {"exact", "similar"}, optional
-            The precision to use when moving the cases. Options are 'exact'
-            or 'similar'. If not specified, "exact" will be used.
-            Ignored if case_indices is specified.
-        preserve_session_data : bool, default False
-            When True, will move cases without cleaning up session data.
-        session : str, optional
-            The identifier of the Trainee session to associate the move with.
-        source_id : str, optional
-            The source trainee unique id from which to move cases. Ignored
-            if source_name_path is specified. If neither source_name_path nor
-            source_id are specified, moves cases from the trainee itself.
-        source_name_path : list of str, optional
-            List of strings specifying the user-friendly path of the child
-            subtrainee from which to move cases.
-        target_name_path : list of str, optional
-            List of strings specifying the user-friendly path of the child
-            subtrainee to move cases to.
-        target_id : str, optional
-            The target trainee id to move the cases to. Ignored if
-            target_name_path is specified. If neither target_name_path nor
-            target_id are specified, moves cases to the trainee itself.
-
-        Returns
-        -------
-        dict
-            A dictionary with key 'count' for the number of moved cases.
-        """
-        result = self._execute(trainee_id, "move_cases", {
-            "target_id": target_id,
-            "case_indices": case_indices,
-            "condition": condition,
-            "condition_session": condition_session,
-            "precision": precision,
-            "num_cases": num_cases,
-            "preserve_session_data": preserve_session_data,
-            "session": session,
-            "distribute_weight_feature": distribute_weight_feature,
-            "source_id": source_id,
-            "source_name_path": source_name_path,
-            "target_name_path": target_name_path
-        })
-        if not result:
-            return {'count': 0}
-        return result
-
-    def remove_cases(
-        self,
-        trainee_id: str,
-        num_cases: int = 1,
-        *,
-        case_indices: Optional[Iterable[Tuple[str, int]]] = None,
-        condition: Optional[Dict[str, Any]] = None,
-        condition_session: Optional[str] = None,
-        distribute_weight_feature: Optional[str] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-    ) -> Dict:
-        """
-        Removes cases from a Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        num_cases : int
-            The number of cases to remove; minimum 1 case must be removed.
-            Ignored if case_indices is specified.
-        case_indices : list of tuples
-            A list of tuples containing session ID and session training index
-            for each case to be removed.
-        condition : dict of str to object, optional
-            The condition map to select the cases to remove that meet all the
-            provided conditions. Ignored if case_indices is specified.
-        condition_session : str, optional
-            If specified, ignores the condition and operates on cases for
-            the specified session id. Ignored if case_indices is specified.
-        distribute_weight_feature : str, optional
-            When specified, will distribute the removed cases' weights
-            from this feature into their neighbors.
-        precision : {"exact", "similar"}, optional
-            The precision to use when moving the cases, defaults to "exact".
-            Ignored if case_indices is specified.
-
-        Returns
-        -------
-        dict
-            A dictionary with key 'count' for the number of removed cases.
-        """
-        result = self._execute(trainee_id, "remove_cases", {
-            "case_indices": case_indices,
-            "condition": condition,
-            "condition_session": condition_session,
-            "precision": precision,
-            "num_cases": num_cases,
-            "distribute_weight_feature": distribute_weight_feature,
-        })
-        if not result:
-            return {'count': 0}
-        return result
-
-    def edit_cases(
-        self,
-        trainee_id: str,
-        feature_values: Optional[Iterable[Any]] = None,
-        *,
-        case_indices: Optional[Iterable[Tuple[str, int]]] = None,
-        condition: Optional[Dict[str, Any]] = None,
-        condition_session: Optional[str] = None,
-        features: Optional[Iterable[str]] = None,
-        num_cases: Optional[int] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-        session: Optional[str] = None
-    ) -> Dict:
-        """
-        Edit feature values for the specified cases.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        feature_values : list of object
-            The feature values to edit the case(s) with. If specified as a list,
-            the order corresponds with the order of the `features` parameter.
-        case_indices : Iterable of Sequence[Union[str, int]], optional
-            Iterable of Sequences containing the session id and index, where
-            index is the original 0-based index of the case as it was trained
-            into the session. This explicitly specifies the cases to edit. When
-            specified, `condition` and `condition_session` are ignored.
-        condition : dict, optional
-            A condition map to select which cases to edit. Ignored when
-            `case_indices` are specified.
-        condition_session : str, optional
-            If specified, ignores the condition and operates on all cases for
-            the specified session.
-        features : iterable of str, optional
-            The names of the features to edit. Corresponds to feature_values.
-        num_cases : int, default None
-            The maximum amount of cases to edit. If not specified, the limit
-            will be k cases if precision is "similar", or no limit if precision
-            is "exact".
-        precision : {"exact", "similar"}, optional
-            The precision to use when moving the cases, defaults to "exact".
-        session : str, optional
-            The identifier of the Trainee session to associate the edit with.
-
-        Returns
-        -------
-        dict
-            A dictionary with key 'count' for the number of modified cases.
-        """
-        result = self._execute(trainee_id, "edit_cases", {
-            "case_indices": case_indices,
-            "condition": condition,
-            "condition_session": condition_session,
-            "features": features,
-            "feature_values": feature_values,
-            "precision": precision,
-            "num_cases": num_cases,
-            "session": session,
-        })
-        if not result:
-            return {'count': 0}
-        return result
 
     def pairwise_distances(
         self,
